pyproject.toml
[tool.poetry]
name = "configurando-o-django-rest-framework"
version = "0.1.0"
description = ""
authors = ["JulioVieiraP <134452045+JulioVieiraP@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.2"
django = "^5.1"
djangorestframework = "^3.15.2"
factory-boy = "^3.3.1"
pytest-django = "^4.8.0"
django-extensions = "^3.2.3"


[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
